[SIPM][SIFISC-18687] : Changement du type de donnée capital libéré + modification comportement autocomplete pour éviter la resaisie du siège après une erreur
git-svn-id: https://svn.etat-de-vaud.ch/SVN_SOCLE/registre/unireg/trunk@120956 cb3efa0a-b8b0-9f43-a4c8-343d8b7e4362

+++ unireg/web/src/main/java/ch/vd/uniregctb/tiers/view/EntrepriseCivilView.java
@@ -1,5 +1,7 @@
 package ch.vd.uniregctb.tiers.view;
 
+import java.math.BigDecimal;
+
 import ch.vd.registre.base.date.RegDate;
 import ch.vd.uniregctb.tiers.MontantMonetaire;
 import ch.vd.uniregctb.type.FormeJuridiqueEntreprise;
@@ -14,7 +16,7 @@ public class EntrepriseCivilView {
 	private Integer numeroOfsSiege;
 	private String nomSiege;
 	private FormeJuridiqueEntreprise formeJuridique;
-	private Integer capitalLibere;
+	private BigDecimal capitalLibere;
 	private String devise;
 	private String numeroIde;
 	private boolean inscriteRC;
@@ -80,11 +82,11 @@ public class EntrepriseCivilView {
 		this.formeJuridique = formeJuridique;
 	}
 
-	public Integer getCapitalLibere() {
+	public BigDecimal getCapitalLibere() {
 		return capitalLibere;
 	}
 
-	public void setCapitalLibere(Integer capitalLibere) {
+	public void setCapitalLibere(BigDecimal capitalLibere) {
 		this.capitalLibere = capitalLibere;
 	}
